+++ NEWS.md
@@ -1,10 +1,11 @@
 # NEWS
 
 ## TODO
-* PPT.ReplaceShapeByGraphic: Find shape using text pattern and replace by image
- 
+* allow wildcards in text matching for replacing?
+
 ## 0.3 (under development)
 
+* PPT.ReplaceShapeByGraphic: Find shape using text pattern and replace by image
 * PPT.ReplaceGraphic: the shapes to operate on (e.g. only text fields) can now be specified.
 * PPT.AddGraphicstoSlideExample rewritten, not compatible with older version any more
 * PPT.FitGraphicInShape2 function to place graphics inside shape